Restructure the repository:
* Introduce README.md in markdown.
* Introduce include and src directories.
* Expect students to put implementation in src/BowlingGame.cpp
  and tests in src/BowlingGameTests.cpp.
* Modernise CMakeLists.txt.
* Add examples and improve the intruction in readme.
