+++ include/BowlingGame.hpp
@@ -3,8 +3,8 @@
 class BowlingGame
 {
 public:
-	void roll(int pins) {}
-	int getScore() {}
+	void roll(int pins);
+	int getScore();
 private:
 };
