+++ src/BowlingGame.cpp
@@ -1,0 +1,4 @@
+#include "BowlingGame.hpp"
+
+void BowlingGame::roll(int pins) {}
+int BowlingGame::getScore(){return 1e4;}
