+++ src/BowlingGameTests.cpp
@@ -1,0 +1,7 @@
+#include "catch.hpp"
+#include "BowlingGame.hpp"
+
+TEST_CASE("failed_succesfully")
+{
+  REQUIRE(true == false);
+}
